+++ docs/src/releasenotes.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 * Added `PYTHON_JULIACALL_HEAP_SIZE_HINT` option to configure initial Julia heap size.
+* `Base.elsize` now defined for `PyArray`.
 
 ## 0.9.24 (2025-01-22)
 * Bug fixes.